Summarized (aka, "group by") Aggregate Pagination
In mongo, to do a "group by" query, you specify a complex _id (an object). To layer on pagination, using the existing `applyPaginationToMongoAggregation`, we should account for `_id` being either a primitive data type or an object. lodash's `isEqual` does this and only adds a few cycles when `_id` is a primitive (when `_id` is an object, chances are good that the total number of "unpaginatedItems" will be low, so a bunch of extra cycles to compare objects in a shorter loop should still be unnoticeable to most users).